Ensure the pool shrink task runs even if maxIdleTime isn't configured but when maxLifeTime is.   Remove extra sleep on connection failure.

+++ btm/src/main/java/bitronix/tm/resource/jdbc/PoolingDataSource.java
@@ -405,11 +405,11 @@ public class PoolingDataSource extends ResourceBean implements DataSource, XARes
 
 	/* management */
 
-    public long getInPoolSize() {
+    public int getInPoolSize() {
         return pool.inPoolSize();
     }
 
-    public long getTotalPoolSize() {
+    public int getTotalPoolSize() {
         return pool.totalPoolSize();
     }
